aws - Add instance name to snapshot description
This commit adds the instance name to the description of a snapshot created in the `process_volume_set` method of the `Snapshot` class in `ec2.py`. The instance name is retrieved from the `Tags` of the resource and appended to the description. This change improves the clarity and usefulness of the snapshot description.